feat: new functionalities for old and new schemas
